Move pixel intensity script into load-all
